test: move tests out of src folder
