+++ test/testUtils/getArcInfo.ts
